requirements.txt
flask==3.0.0
Pillow==10.0.0
numpy==2.2.6
werkzeug==3.0.0
pyserial==3.5
